Generate WSDL for headers defined in the WSDD.
Syntax:
        <parameter type="tns:string" xmlns:tns="http://www.w3.org/2001/XMLSchema" inHeader="true"/>
        <parameter type="tns:string" xmlns:tns="http://www.w3.org/2001/XMLSchema" mode="OUT" outHeader="true"/>

If a Parameter in an operation is marked as a inHeader or outHeader
in the WSDD, generate the correct soap:header element for it in the
binding.  Also add the parts attribute to the soap:body element.

The header must appear in the service method signature.  I am not sure
how I feel about that, as it doesn't seem like this is as important as it would
be for the client stubs.

WSDL2Java already correctly generated the deploy.wsdd and the
Stub and server code.

TODO: add a test case (probably in wsdl/stubheaders) that verifies this.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@257330 13f79535-47bb-0310-9956-ffa450edef68