Do not use eager allocation in the Array deserializer.
This will have a side effect: if the XML claims an array has a size of 4,
but there are only 3 elements provided, Axis will return an array of
size 3.  We will not allocate slots for array positions beyond the last
one we have data for.

Change the array test to reflect this difference.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@257340 13f79535-47bb-0310-9956-ffa450edef68