Fix more regressions from Axis 1.1.
- Always register the xmlType and javaType in the type mapping, even
if they already existed.  This goes along with the "last one wins" phillosophy
of the type mapping registry.  Otherwise only the first entry for java.lang.Object
goes in to the class2Pair map, which means that the 1999 Schema is emitted
in WSDL for this type.

- As a side effect of this, the Map type for HashMap needs to be registered
last in the default type mapping, so it is the preferred class to implement Map.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@259386 13f79535-47bb-0310-9956-ffa450edef68