Add a -j switch to WSDL2Java along with an API and an enhancement to the ant task.
This turns on the JAX-RPC 1.1 compatible default type mapping, which generally
avoids the Axis custom types to prefer mapping to standard Java types.
For instance, the XML Schema types Month, Day and Year are mapped to String
and NonPositiveInteger is mapped to BigInteger.

NOTE: I can't seem to verify that the ant task will work with this syntax:
<wsdl2java url="jaxrpc11types.wsdl" jaxrpc11mappings="yes">

Can someone with ant mojo double check that I updated the task correctly?

TODO: Add a test for the new switch.  Update the WSDL2Java test
case generator to emit the right code when this switch is on.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@259603 13f79535-47bb-0310-9956-ffa450edef68