Remove the jaxrpc 1.1 switch from the WSDL2Java Emitter for Axis 1.2. This switch did some of the right things, but no where did we switch the engine to correctly interpret the XML on the wire, nor did it handle the WSDL generation of the right Schema types.  Basically it did more harm than good.  This is going to have to be done correctly when someone has the time.
Also, fix up the usages of the DefaultSOAPEncodingTypeMapping so that
we use createWithDelegate when we are supposed to and when we have
"use" of encoded, we get *that* type mapping instead of the default Schema
type mapping.

These changes are to prep for trying to fix Axis 1.2 so that it prefers
XML Schema types again (per 1.1) instead of soap encoded types now that
the encoding type mapping is "in front" of the Schema types in the encoding
type mapping.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@260079 13f79535-47bb-0310-9956-ffa450edef68