Add a switch to turn on a preference for XML Schema types: "dotNetSoapEncFix"
This addresses Axis-1467:  http://nagoya.apache.org/jira/browse/AXIS-1467

This is a global switch which sets a static flag in the TypeMappingImpl class
and causes it to skip over SOAP encoded types when looking for the QName
of a Java class.

Updated the reference guide with the new switch.

Fix a problem in the JavaServiceDesc class where the type mapping
used by default (for .JWS files for instance) was not the SOAP encoding
mapping when use=encoded.  For deployed services, we did use the soapenc
types if they were encoded.  If someone explicitly sets a type mapping, we
don't mess around with that.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@260137 13f79535-47bb-0310-9956-ffa450edef68