Fix bug Axis-1900 - WSDL2Java fails on SQL reporting service.   http://issues.apache.org/jira/browse/AXIS-1900
When we check to see if an operations parameter schema can be unwrapped
we did not allow choice elements in the squence.

This does not address the real problem, which is that "wrapped" in the symbol
table is a global for the *service* and not the operation.  If we turn it off,
it should be for the operation only.  This would involve revamping the whole
system, so Axis2 will have to solve this.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@261061 13f79535-47bb-0310-9956-ffa450edef68