Patches from Guillaume:
axis-test.wsdl.marshall.patch :
* correct the test case : I was checking for <base64>[][3] xmltype, but
that map to a byte[][][] javatype
I should test for <base64>[3] only

axis-base64-ser.patch :
* DefaultSOAPEncodingTypeMappingImpl.java : fix copy/paste error
* ArrayDeserializer.java : format changes + comments
* ArraySerializer.java : factor code (constructors) + add a check for
base64 types that remove "[]" from dims if found (further explanation below)
* ArraySerializerFactory.java : factor code (constructors)
* JavaDeployWriter.java : remove the condition for emition of
returnItemType if base64 types

Based on the response message, I noticed that the arrayType was
something like xsd:base64Binary[][3]. Given that xsd:base64Binary is
mapped to byte[], it seems logical that xsd:base64Binary[][3] was mapped
to byte[][][] !
Conclusion : the array serialization is buggy :)

Looking at ArraySerializer, I found that the number of bracket depends
of the number of dimensions of the array item class (~141), while the
componentClass is an array, we append "[]" to dims variable.
Because a base64type is mapped to a byte[], the ArraySerializer logic
goe one step too far and add an extra "[]" (or "* " if SOAP 1.2). To
solve this, a workaround is to remove an extra "[]" if we are
serializing a base64 type.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@261298 13f79535-47bb-0310-9956-ffa450edef68