Fix bug 1955: Axis 1.2rc3 does not write derived types from abstract classes to wsdl.   http://issues.apache.org/jira/browse/AXIS-1955
Patch from Jayachandra (ajayachandra@gmail.com):
I've noticed that in the BeanSerializer.java code when writing out type
information for fields, they are just being written out using the types.writeType()
method this method doesn't have the necessary logic to write out
subTypes of the type we are trying to write out. So I tried to use the
types.writeTypeAndSubTypeForPart() method that has the required functionality.


git-svn-id: https://svn.apache.org/repos/asf/webservices/axis1/trunk/java@261389 13f79535-47bb-0310-9956-ffa450edef68